refactor: remove launch_callback_handler from XBlock
The LTI 1.3 Launch callback handler is removed from the XBlock as the
logic has been moved to its corresponding Django View. The associated
tests are also moved from the XBlock testcase to the Django View
TestCase.
